backend/requirements.txt
bcrypt==5.0.0
bleach==6.3.0
blinker==1.9.0
boto3==1.42.48
botocore==1.42.48
cachelib==0.13.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
Deprecated==1.3.1
dnspython==2.8.0
email-validator==2.1.0
Flask==3.0.0
Flask-Bcrypt==1.0.1
Flask-Caching==2.3.1
Flask-Limiter[redis]==4.1.1
redis==5.2.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.3.1
gunicorn==21.2.0
idna==3.11
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.1.0
limits==5.8.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ordered-set==4.1.0
packaging==26.0
pillow==12.1.0
playwright==1.58.0
pluggy==1.6.0
pyee==13.0.0
Pygments==2.19.2
pytest==7.4.3
pytest-base-url==2.1.0
pytest-html==4.2.0
pytest-metadata==3.1.1
pytest-playwright==0.7.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-slugify==8.0.4
reportlab==4.0.7
requests==2.32.5
resend==2.10.0
rich==13.9.4
s3transfer==0.16.0
sentry-sdk==2.52.0
six==1.17.0
SQLAlchemy==2.0.46
psycopg2-binary==2.9.10
stripe==7.9.0
text-unidecode==1.3
typing_extensions==4.15.0
urllib3==2.6.3
webencodings==0.5.1
Werkzeug==3.0.1
wrapt==2.1.1
WTForms==3.1.1
anthropic>=0.28.0
jsonschema==4.21.1
APScheduler==3.10.4
